Work around Gradle 6 behaviour change
In Gradle 6 the default DuplicateStrategy is FAIL. This reverts to previous
behaviour to avoid an error due to two ktlint rulesets declaring a
RuleSetProvider service which, when merged, results in duplicate paths.
These files have no effect in detekt so the duplication can be ignored.

The duplicate path is "META-INF/services/com.pinterest.ktlint.core.RuleSetProvider"